==> metaG/R_analyses/Taxonomy/01_Filtration.Rmd <==
---
title: "metaG filtration"
author: "SRomano"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Load packages.

```{r setup FALSE}
list.of.packages <- c("dplyr")
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)
if (!requireNamespace("BiocManager", quietly = TRUE)){
    install.packages("BiocManager")}



library(dplyr)
library(ggplot2)
library(RColorBrewer)
source("/Scripts/Pevalence_functions.r")
```

# Load metadata + datasets
```{r}
meta<-readRDS("/metaG/R_analyses/Taxonomy/RDS/meta.rds")
all.g<-readRDS("/metaG/R_analyses/Taxonomy/RDS/all_no0_noUnassigned.rds")
```

# Clen data
## Concordance meta-data
```{r}
dim(meta)
dim(all.g)

# are all samples in meta in the mOTU table?
meta<-meta[order(meta$SampleID),]
all.g<-all.g[, order(names(all.g))]
if((all(meta$SampleID == names(all.g)))){
  print("All SampleID in meta are the same as in the Abund table")
} else {stop("meta and Abund table do not have the same samples")}

```
# Clean the datasets
```{r}
if(length(which(rowSums(all.g)==0)) > 0){
  print("Removing taxa with 0 counts, most likely due to sample filtration")
  message(paste0("There are ", length(which(rowSums(all.g)==0)), " taxa with 0 counts. removed!"))
  all.g<-all.g[-1*which(rowSums(all.g)==0),]
} else {print("No taxa with 0 counts")}

if(length(all.g[, colSums(all.g)==0])  > 0){
  print("Removing samples with 0 counts, most likely due to sample filtration")
  message(paste0("There are ",length(which(colSums(all.g)==0)), " samples with 0 counts. removed!"))
  all.g<-all.g[,-1*which(colSums(all.g)==0)]
} else {print("No samples with 0 counts")}

```

# Filtration by detection
## Load list_df

```{r}
all.g<-readRDS("/metaG/R_analyses/Taxonomy/RDS/list_df_no0_noUnassigned.rds")

print("Convert the data into numeric")
for(i in 1:length(all.g)){
  tax<-row.names(all.g[[i]])
  all.g[[i]]<-apply(all.g[[i]], as.numeric, MARGIN = 2) %>% as.data.frame()
  row.names(all.g[[i]])<-tax
}

print("Remove taxa with 0 and samples with 0")
for(i in 1:length(all.g)){
  message(paste0("There are ", length(which(rowSums(all.g[[i]])==0)), " taxa with 0 counts. removed!"))
  all.g[[i]]<-all.g[[i]][which(rowSums(all.g[[i]])!=0),]
}
for(i in 1:length(all.g)){
  message(paste0("There are ",length(which(colSums(all.g[[i]])==0)), " samples with 0 counts. removed!"))
  all.g[[i]]<-all.g[[i]][,which(colSums(all.g[[i]])!= 0)] 

}

all.g<-lapply(all.g, function(x) x[,which(names(x) %in% meta$SampleID)])


print("Transform in prop")
all.g.prop<-all.g
for(i in 1:length(all.g)){
  all.g.prop[[i]]<-as.data.frame(prop.table(as.matrix(all.g[[i]]), margin = 2))
}

x<-unlist(lapply(all.g, function(x) names(x)))
y<-unlist(lapply(all.g.prop, function(x) names(x)))
all(x==y)

if(all(x[order(x)] == meta$SampleID)){
  print("all Abund table have the same samples than meta")
  saveRDS(all.g, "//metaG/R_analyses/Taxonomy/RDS/l.df_g_counts.rds")
  saveRDS(all.g.prop, "/metaG/R_analyses/Taxonomy/RDS/l.df_g_TSS.rds")
} else {message("Samples in meta and abund table do not match!")}



```

# Calculate and save prevalence
```{r}
l_df<-readRDS("/metaG/R_analyses/Taxonomy/RDS/l.df_g_TSS.rds")

spec_5x10<-prevalence.matching(l_df, n.studies = 2, prev = 5) 
saveRDS(spec_5x10, "/metaG/R_analyses/Taxonomy/RDS/spec_5x10.rds")
spec_10x10<-prevalence.matching(l_df, n.studies = 2, prev = 10)
saveRDS(spec_10x10, "/metaG/R_analyses/Taxonomy/RDS/spec_10x10.rds")
spec_20x10<-prevalence.matching(l_df, n.studies = 2, prev = 20)
saveRDS(spec_20x10, "/metaG/R_analyses/Taxonomy/RDS/spec_20x10.rds")
spec_30x10<-prevalence.matching(l_df, n.studies = 2, prev = 30)
saveRDS(spec_30x10, "/metaG/R_analyses/Taxonomy/RDS/spec_30x10.rds")
```
