==> metaG/R_analyses/Taxonomy/03_DA_metaG.Rmd <==
---
title: "DA metaG"
author: "SRomano"
output: 
  html_document: 
    toc: yes
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r setup, include=FALSE}
list.of.packages <- c("dplyr", "meta", "genodds")
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)
if (!requireNamespace("BiocManager", quietly = TRUE)){
    install.packages("BiocManager")}


library(dplyr)
library(genodds)
library(meta)


set.seed(9821)
```


# Load metadata + datasets
This analysis needs to run on a named list in which each item is the abundance table of a study.

```{r}
meta<-readRDS("/metaG/R_analyses/Taxonomy/RDS/meta.form.rds")
names(meta)<-c("SampleID", "PD", "Other", "Study")
all.g<-readRDS("/metaG/R_analyses/Taxonomy/RDS/all_no0_TSS_noUnassigned.rds")

# order
all.g<-all.g[,order(names(all.g))]
meta<-meta[order(meta$SampleID),]
meta$PD<-factor(meta$PD, levels = c("HC", "PD"))

# are all samples in meta in the OTU table?
stopifnot(all(meta$SampleID == names(all.g)))
message("Are all samples in meta identical to the microbiome table? ", all(meta$SampleID == names(all.g)))

# rename meta for genodds
names(meta)<-c("Sample_ID", "PD", "Other", "Cohort")

# add study to the df and split in a list
# transpose
all.g<-as.data.frame(t(all.g))
stopifnot(all(meta$SampleID == rownames(all.g)))

all.g$study<-meta$Cohort
l_df<-split(all.g, f = as.factor(all.g$study))


# remove the variable study
lapply(l_df, dim)
l_df<-lapply(l_df, function(x){
  x<-x[,-which(names(x) == "study")]
  return(x)
})
```

# Filter
Detection 5% in 2 studies
```{r}
spec<-readRDS("/metaG/R_analyses/Taxonomy/RDS/spec_5x10.rds")

l_df<-lapply(l_df, function(x) x[,which(names(x) %in% spec)])


```

# Calculate DA using GenOdds
```{r, message =F, warning =F, error=FALSE}
# source the function
source("/Scripts/DA_function_gen_odd.R")

df.cross<-calcualte_da_genOdds(list.df = l_df, 
                meta = meta,
                treat = "PD",
                assume_no_effect = TRUE)

saveRDS(df.cross, "metaG/R_analyses/Taxonomy/RDS/df.cross.rds")
```

## Perform meta

```{r}
bins<-unique(df.cross$bin)

df<-data.frame(TE.random = vector(),
                 seTE.random = vector(),
                 lower.random = vector(),
                 upper.random = vector(),
                 pval.random = vector(),
                 bins = vector())

l.met<-vector(mode = "list", length = length(bins))
for(b in 1:length(bins)){
  require(meta)
    tmp.bin<-df.cross[df.cross$bin == bins[b],]
    print(b)
        
    print(dim(tmp.bin))

    x<-try(met<-metagen(TE=estimate, seTE=std.error,data=tmp.bin, method.tau = "REML",title =bins[b], studlab = Cohort))
    if(class(x)[1] != "try-error"){
    met.df<-as.data.frame(c(met[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")],bins= bins[b])) 
    df<-rbind.data.frame(df, met.df)  
    } else { tmp<-data.frame(TE.random =NA, seTE.random = NA, lower.random = NA, upper.random = NA, pval.random = NA, bins= bins[b])
      df<-rbind.data.frame(df, tmp)}
}

df$p.adjust<-p.adjust(df$pval.random, method = "fdr")
```

## Subset sign
```{r}
subset(df,p.adjust<0.05 ) %>%
  dim()
df.sign<-subset(df,p.adjust<0.05 ) 

df.sign$dir<-ifelse(df.sign$TE.random<0, "CTRL", "PD")
saveRDS(df.sign, "/metaG/R_analyses/Taxonomy/RDS/df.sign.rds")

```


# Save workspace
```{r}
save.image("/metaG/R_analyses/Taxonomy//RData/DA.GenOdd.RData")
```
