==> metaG/R_analyses/Taxonomy/08_Coefficients.Rmd <==
---
title: "Coefficients"
author: "SRomano"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r setup}
library(phyloseq)
library(vegan)
library(dplyr)
library(SIAMCAT)
library(tidyverse)

source("/Scripts/Model_coefficient_extraction.r")
set.seed(56987)

```

## Load metadata

```{r}
meta<-readRDS("/metaG/R_analyses/Taxonomy/RDS/meta.rds")
meta$Study<-as.character(meta$Study)
```

## Load LOG models
```{r}
load("/metaG/R_analyses/Taxonomy/RData/metaG_Sinlge_5perc_log.RData")

lridge<-coefficient.extraction(studies = unique(meta$Study) %>% as.character(),
                                list.models = ridge_5x10,
                                model.parallel = T,
                                string.to.add = "log_ridge_5x10",
                                select = "mean.weight")
lridge<-lapply(lridge, function(x) mutate(x, relative.weigths= x$overall.average / sum(abs(x$overall.average))))
lapply(lridge, dim)

df1<-do.call(rbind, lridge) %>% as.data.frame()

```

# Make a figure including taxa freq
```{r}
# add a dummy variable used for formatting
df1$dummy<-rep(1, nrow(df1))
df1$pos.neg<-ifelse(df1$relative.weigths > 0, "POS", "NEG") %>% as.factor()


# make a df with average and std of relative.weights
taxa.weights<-df1 %>%
  group_by(taxa) %>%
  summarise(average = mean(relative.weigths),
            stdev = sd(relative.weigths)) %>%
  as.data.frame()
taxa.weights<-taxa.weights[order(taxa.weights$average, decreasing = T),]
taxa.weights$dir<-ifelse(taxa.weights$average > 0 , "HC", "PD")

saveRDS(taxa.weights, "/metaG/R_analyses/Taxonomy/RDS/Ridge_taxa.weights.rds")

# add average
df.pos.neg<-df1 %>% 
   group_by(taxa, pos.neg) %>%
   summarise(dummy2 = n()) %>%
   as.data.frame()

for(i in 1:nrow(df.pos.neg)){
  df.pos.neg$average[i]<-taxa.weights[taxa.weights$taxa == df.pos.neg$taxa[i], "average"]
  df.pos.neg$stdev[i]<-taxa.weights[taxa.weights$taxa == df.pos.neg$taxa[i], "stdev"]

}
saveRDS(df.pos.neg, "/metaG/R_analyses/Taxonomy/RDS/Ridge_df.pos.neg.rds")

```
