==> metaG/R_analyses/Taxonomy/11_Combinatorics_LOSO_figs.Rmd <==
---
title: "Combinatorics_LOSO_fig"
output: html_document
---
# Load packages
```{r}
library(dplyr)
library(SIAMCAT)
library(ggplot2)
set.seed(234)
```

# Load data
```{r}
bok1<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Boktor_1_2023.rds")
bok1$test<-rep("Boktor_1_2023", nrow(bok1))

bok2<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Boktor_2_2023.rds")
bok2$test<-rep("Boktor_2_2023", nrow(bok2))

bedarf<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Bedarf_2017.rds")
bedarf$test<-rep("Bedarf_2017", nrow(bedarf))

jo<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Jo_2022.rds")
jo$test<-rep("Jo_2022", nrow(jo))


mao<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Mao_2021.rds")
mao$test<-rep("Mao_2021", nrow(mao))


qian<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Qian_2020.rds")
qian$test<-rep("Qian_2020", nrow(qian))


wa<-readRDS("/metaG/R_analyses/Taxonomy/RDS/loso_ridge.combinatorics.AUCs.Wallen_2022.rds")
wa$test<-rep("Wallen_2022", nrow(wa))

# make a single df
df.all<-rbind.data.frame(bok1, bok2, bedarf, jo, mao, qian, wa)
```
# Make a plot
```{r}

(gg<-ggplot(df.all, aes(n.samples, AUC, size = trainings,
                        color = as.factor(trainings))) +
  geom_point() +
  theme_bw() +
  xlab("Number of samples used for training") +
  labs(size = "Number of training\ndatasets",
       color = "Number of training\ndatasets") +
  facet_wrap(~test,
             nrow = 4, dir = "h", scales = "free_x") +
  theme(strip.text = element_text(size = 12),
        axis.text = element_text(size =12),
        axis.title = element_text(size = 14),
        legend.text = element_text(size =12),
        legend.title = element_text(size = 14),
        legend.background = element_blank(),
        legend.position = c(0.825, 0.32),
        legend.direction = "horizontal",
        legend.spacing.y = unit(0.07, 'cm'),
        legend.spacing.x = unit(0.01, "cm"),
        panel.grid = element_blank()) +
   scale_color_viridis_d(option = "viridis"))
```

# Testing
Log transform sample number and standardize everything
```{r}

df.all$log.n.samp<-log10(df.all$n.samples)
df.all$log.scale.n.sam<-scale(df.all$log.n.samp)

mod2<-nlme::lme(AUC ~ log.scale.n.sam, random = ~ 1 | test, data = df.all)
summary(mod2)
nlme::intervals(mod2) # extract 95% ci

#extract var explained
(x<-performance::model_performance(mod2))
(performance::icc(mod2, robust = T))


df.r2<-data.frame(r2 = c(x$ICC, x$R2_marginal),
                  var = c("Test_set", "#Samples"))

(var<-ggplot(df.r2, aes(var, r2, fill = var)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.5) +
  scale_fill_grey() +
  theme_bw() +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.text = element_text(size = 12),
        axis.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.title = element_text(size =14), 
        legend.position = "none") +
    scale_y_continuous(limits = c(0, 0.2), 
                       breaks = c(0,0.025,0.05, 0.075,0.1,0.125,0.15, 0.175,0.2)) +
  xlab("Variable") +
  ylab(expression(R^2)) )

gg + patchwork::inset_element(
  var, 
  left = 0.51, 
  bottom = -0.02, 
  right = 1, 
  top = 0.22
)


ggsave("/metaG/R_analyses/Taxonomy/Figures/Var.expl.combinatorics.jpeg",
       width = 12, height = 12)


```
